+++ app/controllers/api/v1/projects_controller.rb
@@ -1,7 +1,9 @@
 class Api::V1::ProjectsController < Api::V1::BaseController
   def index
-    @bucket = Project
-    @bucket = @bucket.sort_by_name.for_client_id(params[:client_id])
+    @bucket = Project.sort_by_name
+    params[:client_id] = Client.where(:initials => params[:client_initials]).first.id unless params[:client_initials].blank?
+    @bucket = @bucket.for_client_id(params[:client_id]) unless params[:client_id].blank?
+    @bucket = @bucket.where(:git_repo_url => params[:git_repo_url]) unless params[:git_repo_url].blank?
     unless admin?
       @bucket = @bucket.for_user(current_user)
     end