+++ app/controllers/api/v1/tickets_controller.rb
@@ -2,23 +2,42 @@ class Api::V1::TicketsController < Api::V1::BaseController
   def index
     @bucket = Ticket
 
-    @bucket = @bucket.where(:project_id => params[:project_id])
+    @bucket = @bucket.where(:project_id => params[:project_id]) unless params[:project_id].blank?
+    @bucket = @bucket.for_repo_url_and_branch(params[:repo_url],params[:branch]) unless params[:repo_url].blank? && params[:branch].blank?
 
     @tickets = @bucket.order("name").all
-    render :json => @tickets
+    json_array = []
+    @tickets.each do |ticket|
+      json_hash = {
+          :id                   => ticket.id,
+          :name                 => ticket.name,
+          :estimated_hours      => ticket.estimated_hours,
+          :percentage_complete  => ticket.percentage_complete,
+          :hours                => ticket.hours
+      }
+      json_array << json_hash
+    end
+    render :json => json_array.to_json
   end
 
   def show
     @ticket = Ticket.find params[:id]
 
     json_hash = {
-        :name           => @ticket.name,
+        :name                 => @ticket.name,
+        :estimated_hours      => @ticket.estimated_hours,
+        :percentage_complete  => @ticket.percentage_complete,
+        :hours                => @ticket.hours
     }
-    if @ticket.estimated_hours && @ticket.percentage_complete
-      @pie_chart_url = Gchart.pie(:title => "Remaining Estimated Hours vs Hours Worked -- #{@ticket.percentage_complete}% complete",
-        :size => '450x230', :bar_color => "76959c,364447", :data => [100 - @ticket.percentage_complete, @ticket.percentage_complete], :labels => ["Remaining Estimated", "Hours Worked"])
-      json_hash[:pie_chart_url] = @pie_chart_url
-    end
     render :json => json_hash
+  end
+
+  def create
+    @ticket = Ticket.new(params[:ticket])
+    if @ticket.save
+      render :json => {:success => true} and return
+    else
+      render :json => {:success => false, :errors => @ticket.errors.full_messages.to_sentence} and return
+    end
   end
 end