+++ features/step_definitions/devise_steps.rb
@@ -8,11 +8,11 @@ Given /^I am an authenticated user(?: with an? (\w+) role)?$/ do |role|
     @current_user.has_role!(:client, p)
   end
   visit new_user_session_path
-  And %{I fill in "user_email" with "current_user@example.com"}
-  And %{I fill in "user_password" with "password"}
-  And %{I press "Sign in"}
+  step %{I fill in "user_email" with "current_user@example.com"}
+  step %{I fill in "user_password" with "password"}
+  step %{I press "Sign in"}
 end
 
 Then /^I should see the following users:$/ do |expected_users_table|
-   expected_users_table.diff!(tableish('table tr', 'td,th'))
+   expected_users_table.diff!(find('table').all('tr').map { |row| row.all('th, td').map { |cell| cell.text.strip } })
 end